Fix linter issues
