+++ src/execute.c
@@ -6,7 +6,7 @@
 /*   By: gariadno <gariadno@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/28 23:03:34 by aroque            #+#    #+#             */
-/*   Updated: 2021/03/18 23:56:27 by aroque           ###   ########.fr       */
+/*   Updated: 2021/03/19 08:32:24 by aroque           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,6 @@ int		execute_process(t_process *p, t_shell *s, int in, int out)
 	bool	builtin;
 
 	status = 0;
-	builtin = false;
 	redirect_handler(p, in, out);
 	status = execute_builtin(p, s, &builtin, out);
 	if (!builtin)
