+++ src/lexer.c
@@ -6,7 +6,7 @@
 /*   By: aroque <aroque@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/26 21:44:50 by aroque            #+#    #+#             */
-/*   Updated: 2021/03/18 23:22:29 by aroque           ###   ########.fr       */
+/*   Updated: 2021/03/19 08:47:52 by aroque           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,48 +17,7 @@
 #include "minishell.h"
 #include <stdbool.h>
 
-static void	replace_env(char **str, t_hashtable *env, size_t *i)
-{
-	size_t	j;
-	char	*key;
-	char	*value;
-
-	j = 0;
-	value = NULL;
-	if ((*str)[1] == '?')
-		j++;
-	else
-	{
-		while (ft_isalnum_or_uscore((*str)[*i + j + 1]))
-			j++;
-	}
-	key = ft_substr(*str, *i, j + 1);
-	value = get_value(env, key + 1);
-	if (!value)
-		value = "";
-	*str = ft_strreplace(str, key, value);
-	*i += ft_strlen(value);
-	j = 0;
-	free(key);
-}
-
-int			expand_env(char **str, t_hashtable *env)
-{
-	size_t	i;
-
-	i = 0;
-	while ((*str)[i])
-	{
-		if ((*str)[i] == '$')
-			replace_env(str, env, &i);
-		else if ((*str)[i] == '~')
-			ft_strreplace(str, "~", get_value(env, "HOME"));
-		i++;
-	}
-	return (0);
-}
-
-static int	lex_squote(t_token *token)
+static int		lex_squote(t_token *token)
 {
 	size_t	len;
 	char	*tmp;
@@ -73,7 +32,7 @@ static int	lex_squote(t_token *token)
 	return (0);
 }
 
-static int	lex_dquote(t_token *token, t_hashtable *env)
+static int		lex_dquote(t_token *token, t_hashtable *env)
 {
 	size_t	len;
 	char	*tmp;
@@ -89,12 +48,12 @@ static int	lex_dquote(t_token *token, t_hashtable *env)
 	return (0);
 }
 
-bool		special_token(t_token *token, const char *str)
+static bool		special_token(t_token *token, const char *str)
 {
 	return (ft_streq(token->value, str) && token->type == T_UNDEFINED);
 }
 
-int			lexer(t_token *token, t_hashtable *env)
+int				lexer(t_token *token, t_hashtable *env)
 {
 	int		err;
